Remove all the barrier stuff.
I had put it in there to try and control the timing of when the threads began
running, but it was all unnecessary and a bit unrealistic.  I found a bug in
the upthread* schedulers related to when they were requesting vcores, which
appears to be why the data I was seeing was confusing.